typo

+++ src/alchemy/y/alchemy/lib/load.cpp
@@ -15,11 +15,11 @@ namespace upsylon
             lua_settop(L,0);
             lua_getglobal(L,*id);
             if( ! lua_istable(L,-1) ) throw exception("%s%s '%s' is not a table",CLID,fn,*id);
-            const unsigned long n = lua_rawlen(L,-1);
-            for(unsigned long i=1;i<=n;++i)
+            const int n = int(lua_rawlen(L,-1));
+            for(int i=1;i<=n;++i)
             {
                 lua_rawgeti(L,-1,i);
-                if(!lua_isstring(L,-1)) throw exception("%s%s %s[%lu] is not a string",CLID,fn,*id,i);
+                if(!lua_isstring(L,-1)) throw exception("%s%s %s[%d] is not a string",CLID,fn,*id,i);
                 get( lua_tostring(L,-1) );
                 lua_pop(L,1);
             }
